refactor: transfer modal
